requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
marshmallow==3.13.0
flask-marshmallow==0.14.0
SQLAlchemy==1.4.23
Flask-SQLAlchemy==2.5.1
marshmallow-sqlalchemy==0.26.1
PyGithub==1.55
python-dateutil==2.8.2